refactor(categories): :recycle: Complex values in state
Moved computed who return function into state. Moved roulement into state. Removed useless computed values.
