+++ store/categories/getters.ts
@@ -2,7 +2,8 @@ import type { GetterTree } from 'vuex'
 import type { RootState } from '../state'
 import type { CategoryState } from './state'
 import { ECategoryType } from '~/ts/ECategoryType'
-import type { AgendaComputed, Category } from '~/ts/types'
+import type { Account, AgendaComputed, Category } from '~/ts/types'
+import { toLS } from '~/ts/format'
 
 /**
  * Getters for user's categories
@@ -15,7 +16,7 @@ const getter: GetterTree<CategoryState, RootState> = {
    * @param rootGetters The getters of the whole store
    * @returns The categories
    */
-  getCategories: (state, _a, _b, rootGetters) => {
+  getCategories: (state, _getters, _rootState, rootGetters) => {
     const agenda = rootGetters['agenda/getCurrent'] as AgendaComputed
     const weekcount = rootGetters.getWeekCount
 
@@ -50,6 +51,8 @@ const getter: GetterTree<CategoryState, RootState> = {
           break
       }
 
+      const ratio = Math.abs(ba) / Math.abs(b)
+
       categories.push({
         id,
         name,
@@ -57,10 +60,64 @@ const getter: GetterTree<CategoryState, RootState> = {
         balance: ba,
         type,
         icon: i,
+        computed: {
+          // Tooltip for detailed usage of a Category
+          tooltip: `${toLS(Math.abs(ba))} / ${toLS(Math.abs(b))} (${toLS(
+            ratio,
+            {
+              style: 'percent',
+            }
+          )})`,
+          // Calculate usage of a Category and format it
+          usage: toLS(b - ba),
+          // Calculate usage ratio of a Category
+          ratio: {
+            color: ratio < 0.5 ? 'green' : ratio <= 0.75 ? 'orange' : 'red',
+            value: ratio,
+          },
+        },
       })
     }
 
     return categories
+  },
+  /**
+   * Total balance of account minus categories budget
+   */
+  getRoulement: (_state, getters, _rootState, rootGetters): string => {
+    const account = rootGetters['account/getCurrent'] as Account
+    if (account) {
+      let value = account.balance
+      const check = {
+        plannedCredit: false,
+        plannedDebit: false,
+      }
+      for (const {
+        budget,
+        balance,
+        type,
+      } of getters.getCategories as Category[]) {
+        switch (type) {
+          case ECategoryType.PLANNED_CREDIT:
+            value += budget - balance
+            check.plannedCredit = true
+            break
+          case ECategoryType.PLANNED_DEBIT:
+            value += balance - budget
+            check.plannedDebit = true
+            break
+          default:
+            value -= Math.max(budget - balance, 0)
+            break
+        }
+      }
+
+      if (check.plannedCredit && check.plannedDebit) {
+        return toLS(value)
+      }
+      return '-1'
+    }
+    return '-1'
   },
   /**
    * Get the loading state
